celeste-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
huggingface_hub
replicate
transformers